fancy multigrab rendering
fan out all the selected cards as the grabbed object

the one issue with this approach is that even though the cards all
disappear from their droppable area on a multigrab, the positions of the
drop targets doesn't actually change---so you have to drag the card fan
over to a "ghost" location if you're moving things to the right.

i'm not sure it's possible to overcome this without horrible hacks.
rbdnd seems to "remember" the size of all its draggables, so the fact
that many have collapsed to zero-width (by disappearing) doesn't seem to
affect its state.  one good alternative may be to just leave tombstone
markers for the grabbed cards in place (i will probably try that out
next).

on the bright side, react-beautiful-dnd implements a nice affordance
that still makes it clear where the drop will occur.