make sure we store only pure CardBases in PlayArea
when these guys are Cards, their implicit TrumpMeta gets stale if the
trump is ever overridden.  this causes weird incompatibility issues when
we pass them around for various purposes, including sort and play (so,
everything).

because the client state is so transient, it's probably not worth trying
to keep them Card'd correctly, so just strip them to bare CardBases